Measure height

+++ 9-binary_tree_height.c
@@ -20,9 +20,9 @@ int max(int a, int b)
 * Return: None
 */
 
-size_t binary_tree_count(const binary_tree_t *tree)
+int binary_tree_count(const binary_tree_t *tree)
 {
-	size_t right = 0, left = 0;
+	int right = 0, left = 0;
 
 	if (tree == NULL)
 	{
@@ -49,7 +49,7 @@ size_t binary_tree_count(const binary_tree_t *tree)
 
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	size_t size = binary_tree_count(tree) - 1;
+	int size = binary_tree_count(tree) - 1;
 
 	return (size >= 0 ? size : 0);
 }
